Make java.lang.Throwable extend scalanative.runtime.Throwable to allow for Scala Native specific features
